call main() directly; add -g3 flag for debug builds
The main function was called by a function pointer to avoid inlining
of the whole main() function. This was done to use flash space between
interrupt vectors, e.g. for the reset handler. It does not affect STM32
devices though because there is no such space. Furthermore you can not
declare the pointer to main as volatile in a naked function. If you want
to avoid inlining of the main(), simply use __attribute__((noinline)) in
the definition of the main function.